moving api into an api folder and changing paths
